base_movies/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
idna==3.4
PyJWT==1.7.1
python-decouple==3.7
python-environ==0.4.54
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
